requirements.txt
paho-mqtt
pigpio
gpiozero